refactor: merge error.ts and logger.ts into internal.ts for all services
Replace per-service error.ts + logger.ts with a single internal.ts that
exports the error class, toError re-export, and logger instance together.
Reduces auxiliary file count by 18 files while preserving the core/ facade pattern.
