--- src/services/action/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createActionLogger = () => new ServiceLogger("ActionService");
