--- src/services/bookmark/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createBookmarkLogger = () => new ServiceLogger("BookmarkService");
