--- src/services/content/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createContentLogger = () => new ServiceLogger("ContentService");
