--- src/services/history/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createHistoryLogger = () => new ServiceLogger("HistoryService");
