+++ src/services/result/internal.ts
@@ -1,7 +1,10 @@
 import { ServiceError, toError } from "../core/error";
+import { ServiceLogger } from "../core/logger";
 
 export class ResultServiceError extends ServiceError {
   readonly code = "RESULT_SERVICE_ERROR";
 }
 
 export { toError };
+
+export const logger = new ServiceLogger("ResultService");
