--- src/services/result/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createResultLogger = () => new ServiceLogger("ResultService");
