+++ src/services/runtime/internal.ts
@@ -1,7 +1,10 @@
 import { ServiceError, toError } from "../core/error";
+import { ServiceLogger } from "../core/logger";
 
 export class RuntimeServiceError extends ServiceError {
   readonly code = "RUNTIME_SERVICE_ERROR";
 }
 
 export { toError };
+
+export const logger = new ServiceLogger("RuntimeService");
