--- src/services/runtime/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createRuntimeLogger = () => new ServiceLogger("RuntimeService");
