--- src/services/storage/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createStorageLogger = () => new ServiceLogger("StorageService");
