--- src/services/suggestion/logger.ts
@@ -1,4 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createSuggestionLogger = () =>
-  new ServiceLogger("SuggestionService");
