--- src/services/tab/logger.ts
@@ -1,3 +1,0 @@
-import { ServiceLogger } from "../core/logger";
-
-export const createTabLogger = () => new ServiceLogger("TabService");
